End of migration of integration tests
