+++ tests/integration/tests/autocomplete.desktop.js
